+++ extensions/vscode/src/utils/names.ts
